requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via
    #   django
    #   djoser
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
coreapi==2.3.3
    # via djoser
coreschema==0.0.4
    # via coreapi
cryptography==38.0.3
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==3.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.13.0
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.8.0
    # via djoser
djoser==2.1.0
    # via -r requirements.in
idna==3.4
    # via requests
importlib-metadata==1.7.0
    # via djoser
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markupsafe==2.1.1
    # via jinja2
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
pycparser==2.21
    # via cffi
pyjwt==2.6.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pymysql==1.0.2
    # via -r requirements.in
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.6
    # via
    #   django
    #   djangorestframework
requests==2.28.1
    # via
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
six==1.16.0
    # via social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.3.0
    # via social-auth-app-django
sqlparse==0.4.3
    # via django
typing-extensions==4.4.0
    # via asgiref
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.12
    # via requests
zipp==3.10.0
    # via importlib-metadata
